Extracted the taskId criteria while searching the input. #131
The very presence of an "Id" in the name suggests that this entity must
be something unique. But, previously the "taskId" was used among of the
command and its arguments while searching the input object in the
resolveTaskToInput() function. Among of other effects, this was leading
to exception throwing for the use-case like described in the ticket
(combined usage of the Shell Command and CommandVariable extensions
usage). This patch if doesn't fix this issue, then gives a workaround
for it. Specifying the "taskId" for the Shell Command calls lets the
variables substitution with CommandVariable extension (using the
`variableSubstArgs` argument).